add items in user model

+++ server/controllers/item.controller.js
@@ -1,4 +1,5 @@
 const Item = require('../models/Item');
+const jwt = require('jsonwebtoken');
 
 /**
  * create a new item
@@ -7,11 +8,19 @@ const Item = require('../models/Item');
  */
 exports.createItem = (req, res) => {
     delete req.body._id;
+
+    const token = req.headers.authorization.split(' ')[1];
+    const decodedToken = jwt.verify(token, process.env.SECRET_KEY);
+    const userId = decodedToken.userId;
+
     const item = new Item({
         ...req.body,
+        userId: userId,
     });
     item.save(item)
-        .then(() => res.status(201).json({ message: 'Object recorded !' }))
+        .then(() =>
+            res.status(201).json({ message: 'Object recorded !', item })
+        )
         .catch((error) => res.status(400).json({ error }));
 };
 
@@ -22,7 +31,9 @@ exports.createItem = (req, res) => {
  */
 exports.updateItem = (req, res) => {
     Item.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Object updated !' }))
+        .then(() => {
+            res.status(200).json({ message: 'Object updated !' });
+        })
         .catch((error) => res.status(400).json({ error }));
 };
 
@@ -54,7 +65,22 @@ exports.getAllItems = (req, res) => {
  * @param {Express.Response} res
  */
 exports.deleteItem = (req, res) => {
-    Item.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Object deleted !' }))
-        .catch((error) => res.status(400).json({ error }));
+    Item.findOne({ _id: req.params.id })
+        .then((item) => {
+            if (item.userId != req.auth.userId) {
+                res.status(401).json({ message: 'unathorized' });
+            } else {
+                Item.deleteOne({ _id: req.params.id })
+                    .then(() =>
+                        res.status(200).json({ message: 'Object deleted !' })
+                    )
+                    .catch((error) => res.status(400).json({ error }));
+            }
+        })
+        .catch((error) => res.status(401).json({ error }));
 };
+// exports.deleteItem = (req, res) => {
+//     Item.deleteOne({ _id: req.params.id })
+//         .then(() => res.status(200).json({ message: 'Object deleted !' }))
+//         .catch((error) => res.status(400).json({ error }));
+// };
